fixed git source

+++ codev/providers/sources/git.py
@@ -3,14 +3,13 @@ from git import Repo
 
 
 class Git(object):
-    def __init__(self, *args, version=None, commit=None, repository_url=None, directory=None, **kwargs):
-        self.branch = version or None
-        self.tag = version or None
-        self.commit = commit or None
-        self.directory = directory or 'directory'
+    def __init__(self, version=None, commit=None, repository_url=None, directory=None):
+        self.branch = version
+        self.tag = version
+        self.commit = commit
+        self.directory = directory
         self.repository = Repo()
         self.repository_url = repository_url or self.repository.remotes.origin.url
-        super().__init__(*args, **kwargs)
 
     def install(self, performer):
         """
@@ -34,31 +33,43 @@ class Git(object):
 
         performer.execute('mkdir -p ~/.ssh')
         performer.send_file('~/.ssh/known_hosts', '~/.ssh/known_hosts')
-        if performer.check_execute('[ -d {directory} ]'.format(directory=self.directory)):
-            performer.check_execute('rm -rf {directory}'.format(directory=self.directory))
+
+        if self.directory:
+            if performer.check_execute('[ -d {directory} ]'.format(directory=self.directory)):
+                performer.check_execute('rm -rf {directory}'.format(directory=self.directory))
+
+        if self.directory:
+                directory = self.directory
+        else:
+            directory = '.'
 
         if self.branch or self.tag:
+
             performer.execute('git clone {url} --branch {object} --single-branch {directory}'.format(
                 url=self.repository_url,
-                directory=self.directory,
+                directory=directory,
                 object=self.branch or self.tag
             ))
         elif self.commit:
-            performer.execute('git init {directory}'.format(directory=self.directory))
-            with performer.change_directory(self.directory):
+            performer.execute('git init {directory}'.format(directory=directory))
+            with performer.change_directory(directory):
                 performer.execute('git remote add origin {url}'.format(url=self.repository_url))
                 performer.execute('git fetch origin {commit}'.format(commit=self.commit))
                 performer.execute('git reset --hard FETCH_HEAD')
 
 
-class GitSource(Git, Source):
+class GitSource(Source):
     provider_name = 'git'
+
+    def __init__(self, *args, **kwargs):
+        self.git = Git(directory='directory')
+        super().__init__(*args, **kwargs)
 
     def process_options(self, options):
         if not options:
             raise ValueError('Repository options must be specified.')
 
-        remote = self.repository.remote()
+        remote = self.git.repository.remote()
         remote.fetch()
 
         # branch
@@ -66,12 +77,12 @@ class GitSource(Git, Source):
             self.branch = options
 
         # tag
-        elif options in self.repository.tags:
+        elif options in self.git.repository.tags:
             self.tag = options
 
         # commit
         else:
-            for commit in self.repository.iter_commits():
+            for commit in self.git.repository.iter_commits():
                 if options == commit:
                     self.commit = commit
             else:
